backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
openai>=1.3.0
ultralytics>=8.0.196
numpy>=1.24.3
opencv-python>=4.8.0
torch>=2.0.0
python-dotenv>=1.0.0
click>=8.0.0
starlette>=0.27.0
pydantic>=2.0.0
typing-extensions>=4.0.0
pillow>=10.0.0 